Rewrite Playfair for v 3.0
Playfair has languished for so long because the code as it was
written was fragile and unmaintainable. It was an early project
for me and I was a novice both at Prolog and as a programmer in
general. It was a good start, but it's long needed modernising.

My first idea was to update the syntax and some high-level stuff
without getting into the weeds of the logic. This became difficult
to do, and as a result this turned into a near complete rewrite.
The code is still a little rough, but it's more readable and
maintainable now. The Playfair-style script code is now separate
from the HTML code, allowing (with any luck) scripts to be parsed
in either direction and for new formats to be added in the future.
The Prolog is now ISO compliant, so it should run with any
interpreter.

The input/output scripts should be exactly the same as in v2.
Playfair supports the same features and should produce the same
HTML output (albeit with different indentation). I'm not yet sure
what else it needs in terms of features, but future work should
be significantly easier now that the code is essentially sane.